feat: separate every ViewModel Object  in a ViewModel.swift
