+++ STUDYA/STUDYA/ViewModel/SignInViewModel.swift
@@ -1,0 +1,40 @@
+//
+//  SignInViewModel.swift
+//  STUDYA
+//
+//  Created by 서동운 on 3/6/23.
+//
+
+import Foundation
+
+// MARK: - SignInViewModel
+
+class SignInViewModel: ViewModel {
+    typealias Model = Credential
+    
+    var credential: Credential {
+        didSet {
+            handler?(credential)
+        }
+    }
+    var handler: DataHandler?
+    
+    init(credential: Credential = Credential()) {
+        self.credential = credential
+    }
+    
+    func singIn(_ successHandler: @escaping (User) -> Void, _ failureHandler: @escaping (PeoplesError) -> Void) {
+        guard let email = credential.email, let password = credential.password else { return }
+        
+        Network.shared.signIn(id: email, pw: password) { result in
+            switch result {
+            case .success(let user):
+                
+                successHandler(user)
+            case .failure(let error):
+                
+                failureHandler(error)
+            }
+        }
+    }
+}
