+++ STUDYA/STUDYA/ViewModel/StudyViewModel.swift
@@ -1,0 +1,80 @@
+//
+//  StudyViewModel.swift
+//  STUDYA
+//
+//  Created by 서동운 on 3/6/23.
+//
+
+import Foundation
+
+class StudyViewModel: ViewModel {
+    typealias Model = Study
+    
+    var study: Study {
+        didSet {
+            handler?(study)
+        }
+    }
+    
+    var handler: DataHandler?
+    
+    init(study: Study = Study()) {
+        self.study = study
+    }
+    
+    func bind(_ handler: DataHandler?) {
+        self.handler = handler
+        handler?(study)
+    }
+    
+    func getStudyInfo() {
+        guard let studyID = study.id else { return }
+        
+        Network.shared.getStudy(studyID: studyID) { [self] result in
+            switch result {
+            case .success(let studyOverall):
+                study = studyOverall.study
+                study.ownerNickname = studyOverall.ownerNickname
+            case .failure(let failure):
+                print(failure)
+            }
+        }
+    }
+    
+    func postNewStudy(_ successHandler: @escaping () -> Void) {
+        Network.shared.createStudy(study) { result in
+            switch result {
+            case .success:
+                successHandler()
+            case .failure(let error):
+                print(error)
+            }
+        }
+    }
+    
+    func updateStudy(successHandler: @escaping () -> Void) {
+        guard let studyID = study.id else { return }
+        
+        Network.shared.updateStudy(study, id: studyID) { result in
+            switch result {
+            case .success:
+               successHandler()
+            case .failure(let failure):
+                print(failure)
+            }
+        }
+    }
+    
+    func deleteStudy(successHandler: @escaping () -> Void) {
+        guard let studyID = study.id else { return }
+        
+        Network.shared.closeStudy(studyID) { result in
+            switch result {
+            case .success:
+                successHandler()
+            case .failure(let error):
+                print(error)
+            }
+        }
+    }
+}
